Add files via upload
The dbcon file is for the connection with the MySQL server. The register.html file is the main page that is the homepage and the register1.html file is for the new users to log in. The PHP files are there to retrieve and insert data into the database. You can run this code by cloning my repository in your local system. I hope that this repository may help you to understand deeply the login systems used by the tech giants.